Refactor `kubernetes_client` to `kubernetes`
Also renames `Kubernetes` to `Client`, so now we can use
`kubernetes.Client` instead of `kubernetes_client.Kubernetes`.
